refactor: 更新proto及客户端导入路径，调整Docker构建配置
